Bug fix: - reduced the visibility to the player of an AI cheat in the   pathfinding algorithm.
The AI cheats by "knowing" the terrain types and movement costs of unexplored
terrain. This is visible to the human player when dragging the cursor into
unexplored territory. The color of the path could be used to obtain
information about the nature of the terrain.

This bug fix will only correct the color of the path. It does not remove the
AI cheat, and you may still be able to detect this by examining the form of
the path.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@22 34d2fc0c-33df-0310-8a34-85f8040ba873