Bug fix: - use the veteran status in combat.
Based on the observations and proposals by MrBaggins and Solver:
- apply the veteran coefficient to the unit properties when in combat.

This bug fix will improve the chance to hit the target (i.e. increase the
RangedAttack and Offense properties).
It will not modify the other properties (Defense, Strength, etc.).
This corresponds to the handling of the veteran status in UnitData.cpp.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@24 34d2fc0c-33df-0310-8a34-85f8040ba873