2004.05.22 ..\ctp2_code\gs\gameobj\CityEvent.cpp ..\ctp2_data\default\gamedata\script.slc ..\ctp2_data\XXXXXX\gamedata\info_str.txt
Where XXXXXX is english, german and french, the needed strings for the English
and German version are added to the according info_str.txt, there are also
strings added to the French version of the info_str.txt unfortunatly in
English in the lack of a French translation. The Italian, Japanese and Spanish
version of the info_str.txt are missing. And therefore even English text for
these versions of CTP2 are missing.

CTP1 reimplementation by Martin Gühmann: During conquest there is now a
chanche of 0.5 accoriding to our const.txt that the attacker takes an advance,
given the victim has any advances that the attacker does not have. Both the
attacker and the victim get a message about the stolen advance.


2004.05.23
..\ctp2_code\gs\dbgen\RecordDescription.cpp
..\ctp2_code\gs\newdb\*Record.cpp
..\ctp2_code\gs\newdb\*Record.h
..\ctp2_code\gs\newdb\*Record.stamp

The * stands as usual for very combination of letters, meaning every in that
directory with Record in its name is included so that no one has to build a
debug version before the release version.

The pupose of this is to stop .NET to complain about the fact that all these
Init functions in these files don't have a reurn type, the return type void
was added to that function. Odd that dbgen did not generated all of the new
Record files, well some hand editing solved that problem.



git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@147 34d2fc0c-33df-0310-8a34-85f8040ba873