- Bug fixes:
PlayerEvent.cpp:
- Corrected the index of the 'cause' argument in the GrantAdvanceEvent
  handler.

SlicFunc.cpp:
- Added input checks and an optional 'cause' argument to the GrantAdvance
  SLIC handler.
- Added reading of the 'army' argument of the ArmyIsValid SLIC handler,
  instead of returning 0 always - even when the army would have been valid.

SlicStack.cpp:
- Prevented crashes that may occur when SLIC argument values are invalid.

Notes for SLIC writers:

The crashes could occur with seemingly harmless constructs like:

code:if (IsValid(x) && AssumeValidUsage(x))


When using C/C++, the AssumeValidUsage(x) part would not have been evaluated
when the IsValid(x) part is false, but with SLIC, the AssumeValidUsage(x) part
is evaluated regardless of the IsValid(x) result.

With the new SlickStack.cpp code, invalid usage should not cause a CtD
anymore. There is a catch, however: invalid expressions may now return 0. When
guarded as in the code example above, this is not a problem - after full
evaluation, the expression in the if will become false. When unguarded, as in
a

code:if (AssumeValidUsage(x) == 0)

naive test to find out whether the Barbarian player is involved, the result
may not be what you expect.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@149 34d2fc0c-33df-0310-8a34-85f8040ba873