2004.05.22 ..\ctp2_code\gs\gameobj\CityEvent.cpp ..\ctp2_data\default\gamedata\script.slc ..\ctp2_data\XXXXXX\gamedata\info_str.txt
Where XXXXXX is english, german and french, the needed strings for the English
and German version are added to the according info_str.txt, there are also
strings added to the French version of the info_str.txt unfortunatly in
English in the lack of a French translation. The Italian, Japanese and Spanish
version of the info_str.txt are missing. And therefore even English text for
these versions of CTP2 are missing.

CTP1 reimplementation by Martin Gühmann: During conquest there is now a
chanche of 0.5 accoriding to our const.txt that the attacker takes an advance,
given the victim has any advances that the attacker does not have. Both the
attacker and the victim get a message about the stolen advance.


2004.06.01
..\ctp2_data\default\uidata\credits.txt

The credits from June the first, by Locutus.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@157 34d2fc0c-33df-0310-8a34-85f8040ba873