Standardization of the behaviour for selecting and focus on a city.
Added the missing focus behaviour when selecting a city for the following
cases:

1) CityControlPanel: next & prev button and the dropdown selection
2) CityWindow: dropdown selection
3) BuildManager: next & prev button and the dropdown selection

Also took care, that the user option "focus on the active unit or city" works
as expected.

The focus handling from the national manager, to select and focus a city with
a double mouse click, was not changed. Because considering here, the user
option not to focus, makes no sense.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@161 34d2fc0c-33df-0310-8a34-85f8040ba873