Bug fix: make sure that the human player will get the selected civilisation.
Notes:Tested for most SP start possibilities. Not tested for MP. The code for
an editor start has not been changed (still using player 1). I could not fully
test it for the Alexander scenario, because it crashes (in a DPRINTF statement
) in InitSStateEvent upon game start. Could you have a look at that one,
Martin? I suspect the crash is caused by the 2004.08.15 personality changes.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@209 34d2fc0c-33df-0310-8a34-85f8040ba873