Backwards compartibilty restoration by Martin Gühmann: If the new structs are missing from personalities.txt the old strategies and priorities are taken as default values. The following startegies data must be defined in block if block member is missing default values are also used for the other members as well:
DefenceNoneStrategy
DefenceVeryLowStrategy
DefenceLowStrategy
DefenceMediumStrategy
DefenceHighStrategy
DefenceVeryHighStrategy

The same is true for the following nuclear strategies:

MinimumNukesStrategy
LowNukesStrategy
AverageNukesStrategy
MaximumNukesStrategy

If the following nuclear strike strategy data are present then the according
nuclear strike strategy is used if it is not present no nuclear strike
strategy is used:

NuclearStrikeStrategy

If the two many cities strategy data are not present, no strategy is loaded to
reduce settling and conquering activity. In that case the no revolution
strategy isn't loaded, either. Even if it is present. If the no revolution
strategy data are missing, then no no revolution startegy is loaded if the
according civ is over the city limit and owns the wonder that makes this civ
to get revolting cities.



git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@216 34d2fc0c-33df-0310-8a34-85f8040ba873