..\ctp2_code\ui\aui_ctp2\linegraph.cpp
Improved .NET vs. VC6 compatibility by Martin Gühmann, replaced casts to long
by long numeric literals.

..\ctp2_data\default\gamedata\Const.txt
..\ctp2_data\default\gamedata\terrain.txt
..\ctp2_data\default\gamedata\tileimp.txt

Set MAX_MATCH_LIST_CYCLES to 6 to accelerate the game.
Devieded CARAVAN_COEF so that it is now 0.004.
Used Flink's tileimp.txt and terrain.txt as base for better Freight values and
multiplyed all the Freight values in both files by ten, so that the devisin by
ten of the CARAVAN_COEF is canceled. In addition the Freight values for rivers
are identical to those of roads, so that it is not more likely that a trade
route follows a river than a road.



git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@220 34d2fc0c-33df-0310-8a34-85f8040ba873