Text file update by Martin Gühmann:
..\ctp2_data\default\terrain.txt
Updated the terrain scores so that match they more to the ones from Apolyton,
this results in a more aggressive settling behaviour of the AI.

..\ctp2_data\default\civilisation.txt
..\ctp2_data\english\civ_str.txt
..\ctp2_data\german\civ_str.txt
..\ctp2_data\italian\civ_str.txt

Reinserted the civs from the original game. And removed most of the new civs
to fit to the UI limitation. The following civs have been readded:

Korean
Nigerian
Welsh
Cuban
Nicaraguan
Irish

The following civs have been removed in comparision to the last version:

Byzantine
Babylonian
Carthaginian
Austrian
Swedish
Celtic

The following civs are now in but are not in the original game.

Arabian
Nubian
Aztec

All the aditional city names and civs names are still in the string files, the
additional civs over the limit of 45 are still in the civilisation.txt but
outcommented to make sure that all civs are selectable by the human player
that are in the game. The downside of this restoration is that the readded
civs just have 30 city names. That is also true for the Nubians and the Poles.
Unfortunatly there is no French and Spanish version of the civ_str.txt
therefore I wasn't able to provide for these languages at least non-translated
strings in English, the consequence of these missing strings are that the
Frensh and Spanish version of CTP2 doesn't run.


git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@225 34d2fc0c-33df-0310-8a34-85f8040ba873