Moved TREATY_TO_EXPIRE from dip2_str.txt to info_str.txt were it was in CTP1. This holds for the English version of these files. For the other languages it was moved directly from cut_str.txt to info_str.txt.
The other diplomatic messeages in cut_str.txt were moved and updated to CTP2 format as well for future reimplementation.

The Spanish cut_str.txt was taken from one of the others probably they are identiacal anyway. However the rest of the Spanish string files are still missing.

Translations are missing as well for the French, Italian, Spanish and Japanese version. Since these strings are from CTP1 it shouldn't be such a big deal to get translations of those. Except for the Japanese version. Except of two files all the files are missing.

git-svn-id: http://ctp2.darkdust.net/anonsvn/trunk@299 34d2fc0c-33df-0310-8a34-85f8040ba873