Make VS 2017 compile and run the code (VS 2008 does not compile)
"array[-1]" should behave in an undefined way. Problem showed up here first:
..\ctp2_code\gs\database\profileDB.cpp

"#warning" does not exist in VS 2017 and for what I could find not in the
standard, either:
..\ctp2_code\gs\fileio\Token.cpp

"Uint32" is not defined only "uint32"
..\ctp2_code\ctp\civ3_main.cpp

Do not comment out, use the preprocessor:
..\ctp2_code\gs\world\wldgen.cpp

Missing #ifdef __linux__; Maybe this should go to the function CI_FixName() to
not clutter the main code:
..\ctp2_code\ctp\ctp2_utils\c3files.cpp
..\ctp2_code\ui\aui_common\aui_bitmapfont.cpp
..\ctp2_code\ui\aui_common\aui_image.cpp
..\ctp2_code\ui\aui_common\aui_sound.cpp

Use the preprocessor to use anything SDL related. No need for
SoundManager::PlaySound, the SDK version is used anyway. Therefore, the thing
is called StupidPlaySound:
..\ctp2_code\sound\soundmanager.cpp
..\ctp2_code\sound\soundmanager.h
..\ctp2_code\ui\aui_common\aui_mouse.cpp

Added "#include <algorithm>" for min/max support, is not compatible with VS6:
..\ctp2_code\gfx\gfx_utils\ColorSet.cpp
..\ctp2_code\gs\gameobj\Exclusions.cpp
..\ctp2_code\ui\netshell\netfunc.cpp
..\ctp2_code\ui\netshell\ns_improvements.cpp
..\ctp2_code\ui\netshell\ns_units.cpp
..\ctp2_code\ui\netshell\ns_wonders.cpp

VS 2017 has stdint.h, breaks compiling on VS 2008:
..\ctp2_code\gfx\gfx_utils\pixeltypes.h

Raise Direct Input to version 8. Version 7 is not included in the Windows SDK:
..\ctp2_code\os\include\config_win32.h
..\ctp2_code\ui\aui_directx\aui_directinput.h
..\ctp2_code\ui\aui_directx\aui_directjoystick.cpp
..\ctp2_code\ui\aui_directx\aui_directjoystick.h
..\ctp2_code\ui\aui_directx\aui_directx.cpp
..\ctp2_code\ui\aui_directx\aui_directx.h

Removed reftime.h, it is not included in DirectX anymore, just added the used
definitions:
..\ctp2_code\ui\aui_common\aui.h
..\ctp2_code\ui\aui_ctp2\directvideo.cpp
..\ctp2_code\ui\aui_directx\aui_directmovie.cpp
..\ctp2_code\ui\aui_directx\aui_directui.h

The code is not used. For now, the code is commented out, but should be
removed:
..\ctp2_code\ui\aui_utils\vidplay.cpp
..\ctp2_code\ui\aui_utils\vidplay.h