Make the AI not to settle too close, and /debugcells fix, fix a crash
Remove any settle goals the AI does not want to settle anymore. Makes the AI not to settle next to a city collection border.
..\ctp2_code\ai\ctpai.cpp

The AI starts considering settle goals when a settler is almost finished
..\ctp2_code\ai\mapanalysis\settlemap.cpp
..\ctp2_code\ai\mapanalysis\settlemap.h
..\ctp2_code\gs\gameobj\BldQue.h
..\ctp2_code\gs\gameobj\bldque.cpp

Fixed a crash with removing goals
..\ctp2_code\ai\strategy\scheduler\Scheduler.h
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp

Fixed /debugcells command and make the game compile again in debug mode
..\ctp2_code\ctp\ctp2_utils\c3cmdline.cpp
..\ctp2_code\ui\interface\chatbox.cpp