Accelerate the AI and reduce the amount of needed memory
New goals are only added if the AI can achieve them. This saves memory, since memory allocation is time intensive, it accelerates the game and since the game is still a 32 bit program, we won't the 2 GB limit so quickly. This makes the late game more stable.
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp

Add goals for armies and cities on each turn. Instead when, cities are captured or destroyed, or armies are created, or on reload. This way, we do not miss anything and we can just add the goals that can also be executed by the AI according to the game state for instance diplomacy.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\ctpai.h

Refactored Get_Totally_Complete and renamed to IsTotallyComplete. IsTotallyComplete is now divided into IsComplete and IsCurrentlyUnavailable. IsComplete is divided into IsInvalid, IsTargetImmune, IsInvalidByDiplomacy. A goal IsInvalid if the target does not anymore exist, for instance a city was destroyed. A goal IsTargetImmune, if its order cannot be used on its target for instance the order enslave cannot be used on units that are not settlers. But of course, if we have the abolish slavery wonder then the order enslave cannot be used either. A goal is IsInvalidByDiplomacy if for instance we have a piece treaty and want to siege a city of that civ. A goal IsCurrentlyUnavailable, if we have a watchful city or not enough gold for paying a spy to incite a revolution, this could change within a few turn.

The AI explores sites if there is a city or an army. Slavery goals are now complete if slavery was abolished. The AI only incites revolutions on empires with more than one city. Those one city civs are so weak that they could be just conquered, since you should have enough resources to do so if you can incite a revolution there.
..\ctp2_code\ai\strategy\goals\Goal.cpp

Corrected the German translation of slave raid.
..\ctp2_data\german\gamedata\ldl_str.txt

Improved the text about the forbidden city. Probably the original English can also be improved.
..\ctp2_data\german\gamedata\Great_Library.txt

Default debug player is now player 1.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\ctpaidebug.cpp

Consecutive changes:
..\ctp2_code\ai\strategy\agents\agent.cpp
..\ctp2_code\ai\strategy\goals\Goal.h
..\ctp2_code\gs\gameobj\ArmyData.cpp
..\ctp2_code\gs\gameobj\ArmyData.h
..\ctp2_code\gs\gameobj\Player.cpp
..\ctp2_code\gs\gameobj\player.h
..\ctp2_code\net\general\net_army.cpp
..\ctp2_code\net\general\net_unit.cpp

White space and comments:
..\ctp2_code\ai\diplomacy\diplomat.cpp
..\ctp2_code\ai\strategy\scheduler\Scheduler.h