Merge the aui_movie from the Linux branch into master
This is the movie manager from the Linux branch, so far I only tested the intro movie, it does not play, I just get a mouse pointer on a black screen. But it least it seems to try to play the movie. I haven't tried the wonder movies, yet. Since, it is part of a bigger chunk, I do not know whether it compiles on its own, it may do. However, I submit it separately, because it would get lost compared to what is to come.
..\ctp2_code\ui\aui_common\aui_movie.cpp
..\ctp2_code\ui\aui_common\aui_movie.h