Merge from the Linux branch aui_music
I think this is supposed to make the music play in the game, however it only plays on Windows but not on  Linux, at least what is concerning my CD. I submit this separately, since there is much more coming and this would get lost in the vast amount of merging. Judging from the code, I think it would compile alone.
..\ctp2_code\ui\aui_common\aui_music.cpp
..\ctp2_code\ui\aui_common\aui_music.h