Enable the logs on Linux and make them run without error
Enable the logs and the memory leak reporter. The logs are now working, they can be found in a subdirectory of executable: ./logs. The memory leak reporter does not work, but it is already creating the leak report files, which can be found in the executable directory: CTP_LEAKS_99999.TXT, CTP_LEAKS_ALT_99999.TXT. So far these files do not contain any usable information, yet.
../ctp2_code/ctp/civ3_main.cpp

The logger tried to delete the directories logs/.. and logs/., which of course it should not try.
../ctp2_code/ctp/ctp2_utils/c3debug.cpp

Fixed a crash that occurred when creating the logs. And increased the buffer size for the debug assert messages, to avoid a stack overflow by writing in unallocated memory.
../ctp2_code/ctp/debugtools/log.cpp