Make the debug system loading the symbols for the functions on Linux
At first it felt a bit hacky to extract the symbols from the executable with nm. But afterwards it doesn't look so bad. However, the call stack and the leak reporter still do not report any useful information. I just can say load CTP2 and then exit does not leak, however if you load a save game, it leaks, I just cannot say what.

CTP2 loads the symbols from the executable
..\ctp2_code\ctp\debugtools\debugcallstack.cpp

Modified so that the function Os::GetExeName() can be accessed from outside of civ3_main.cpp, so that the symbols can be extracted from the executable.
..\ctp2_code\ctp\civ3_main.cpp
..\ctp2_code\ctp\civ3_main.h