Update the save game converter for pull request #89
Another incompatibility was found for Windows savegames loaded on 64bit versions of CTP2, which includes the current Linux version but no Windows version, because so far all Windows versions are 32 bit. It is a problem that only occurs if the player has messages stores, and it only occurs with a subset of messages. Therefore, save games could be converted by saving them without any stored messages. But of course for more convenience here is an update for the save game converter. It loads save games in the old 64 bit format and saves them in the format generated by 32 bits.
..\ctp2_code\gs\slic\SlicArray.cpp