Add a menu item so that the chat Window can be opened without the keyboard
Since the keyboard handling is a bit buggy on Linux so that I cannot use the apostrophe key to open the chat. I need an alternative way to open it. Therefore, I added it to the options menu. It is not the best place, but we do not have a better place. This is also useful for the gamer since it displays the command for the chat window in the options menu.
..\ctp2_code\ui\interface\controlpanelwindow.cpp