Implement auto turn cycling via the chat window for Linux
Enable the the '/rnd x' command in the chat window to run the game for x turns. Press Esc to stop auto turn cycling.
..\ctp2_code\ui\interface\chatbox.cpp

Expose the SDLMessageHandler function so that it can be used in chatbox.cpp. This is necessary so that it can catch the Esc key, while everything else is still handled by the key handler.
..\ctp2_code\ctp\civ3_main.cpp
..\ctp2_code\ctp\civ3_main.h