Clean up white space and minor optimizations
White space only
..\ctp2_code\ctp\civ3_main.cpp

White space only, but including one in the /rnd feedback message, which is strictly a functional change.
..\ctp2_code\ui\interface\chatbox.cpp

White space and minor code optimizations of the type:

	sint32 i;
	for(i=0;i<CP_TILEIMP_MAX;i++)

to

	for(sint32 i = 0; i < CP_TILEIMP_MAX; i++)

Note this makes a difference for Visual Studio 6, as there in both versions i is visible outside of the for loop. However, the code does not compile under VS6, anymore anyway.
..\ctp2_code\ui\interface\controlpanelwindow.cpp