Adjust the save game converter for bcfccf7 in #89
This reflect the last change for the save game format in pull request #89. Here is the message from that commit:

Make Windows savegames with messages load correctly on Linux

The previous commit made save games saved on Windows with certain messages load on Linux. However, it did not load the data, properly. That is because:

archive.Load((uint8 *)m_array, m_arraySize * sizeof(SlicStackValue::m_int));

loads an array, each member is 32 bit on Windows and on Linux 64 bit. The code above does only fill half of the array on Linux, and squeezes into pairs, one member is ordinary and the other is stored in dead space. This leads to an index shift for each second member.

..\ctp2_code\gs\slic\SlicArray.cpp