Make the game not crash on reload, if CTP2 compiled by GCC with optimizations
If CTP2 is compiled by GCC and optimization flags are used such as -Os or -Og, GCC removes the NULL assignments from the destructors. Since the memory for the Order objects comes from a pool, these parts of memory can be reused and the destructor for that is called again. That is the design, whether it is good or bad is another question. However, pointers in this design must be NULLed for proper code function. So for the Order destructor optimizations are disabled.
..\ctp2_code\gs\gameobj\Order.h