Make the AI stop using resources for research if it cannot research anything
This applies to the AI, but also for human players who use the governors and the optimize slider buttons from the user interface, because these use the logic of the AI. In fact the optimize slider buttons are debug tools.

The AI sets the research tax to zero if it cannot research anything
..\ctp2_code\ai\ctpai.cpp

The AI does not use scientists if it cannot research anything. This stops also the governors and the optimize slider buttons to use scientists.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\ai\CityManagement\governor.h
..\ctp2_code\gs\gameobj\Advances.cpp
..\ctp2_code\gs\gameobj\Advances.h
..\ctp2_code\ui\interface\CauseAndEffectTab.cpp
..\ctp2_code\ui\interface\citywindow.cpp

Added the database indices in comments for the advances for better debugging
..\ctp2_data\default\gamedata\Advance.txt