make the AI build the Gaia Controller
Add a build list for the endgame. And add the indices to comments for easier debugging.
..\ctp2_data\default\aidata\BuildingBuildLists.txt

Add the new build list to the build list sequences
..\ctp2_data\default\aidata\buildlistsequences.txt

Also add the indices in comments for better debugging, which was needed here, to identify the computer cores and the sattelites
..\ctp2_data\default\gamedata\buildings.txt

Keep the AI from building more Gaia computers and satellites if it already has enough
..\ctp2_code\ai\CityManagement\governor.cpp
..\ctp2_code\gs\gameobj\gaiacontroller.h