Stop the AI from settling if it clears orders and is not at the target position
Due to the asynchronous execution of events, the move events are executed before the a settle event is added. If during move the AI reveals foreign units it clears the orders. However, the settle order has not been added to the order queue, and without the move orders before the settle order, the AI settles at the place where it cleared orders.
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp
..\ctp2_code\gs\gameobj\ArmyData.cpp
..\ctp2_code\gs\gameobj\ArmyData.h