Speed up event logging
I used event logging to fix the AI settling from the last commit. Therefore I put this into this pull request even so it could go into its own pull request.

The event log is written to a file, which was opened and closed at every write access. Opening and closing means that it allocates each time memory, which is slow. Therefore, I changed it so that the event log files is opened at the beginning and closed at the end of the session. The event log can be enabled by passing the command line option eventlog to the CTP2 executable.
..\ctp2_code\gs\events\GameEventManager.cpp
..\ctp2_code\gs\events\GameEventManager.h