Improve AI debugging by recording also the position type of the army in question
When the decides which army it uses for a goal the debug and logging version writes an entry about the goals raw priority. So far it only recorded the described the target position as either in field or the city. This adds also this bit of information to the position of the army.
..\ctp2_code\ai\strategy\goals\Goal.cpp