Make the strength estimated the same if the army is being transported or not
For computing the needed troop flow the AI calculates how strong an army is, right now it considers the army and its cargo. However, for fighting the strength of the transporter is irrelevant, so only the cargo should be considered and for garrisons the strength of the cargo is irrelevant only the strength of the transporter matters. This has been fixed so that the AI won't put units onto an transporter just to unload them on the next turn to load them again when they are on land.
..\ctp2_code\ai\ctpai.cpp
..\ctp2_code\ai\strategy\agents\agent.cpp
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp
..\ctp2_code\ai\strategy\squads\squad_Strength.cpp