Figure out if a position is in the vision range of a unit correctly
The code fixed is supposed to figure out whether a position is in vision range of a unit. However, the distance to the location was given squared while the vision range was not squared. This is now fixed. This code is used to calculate goal priorities and determines whether a goody hut is in vision range.
..\ctp2_code\gs\gameobj\Unit.cpp