Make the AI use the garrison unit counts from the threat level strategies
The strategies STRATEGY_CAREFUL_START and STRATEGY_AGRESSIVE_START included garrison unit counts even so for the start there are also the defense level strategies, which also set the garrison unit count, but that was overwritten by the start strategies, even so it should not.

Therefore, the start strategies do not define the garrison settings anymore. So that the AI can select the garrison unit count for start phase on the game on threat level.
..\ctp2_data\default\aidata\strategies.txt