Make the AI using the units assigned for garrison for the goal GOAL_DEFEND
At each turn the AI assigns units to the garrisons of cities. It can do whatever it wants with other units such as exploring. However, it should also use the garrison units for the goal GOAL_DEFEND instead of other units. If it uses other units that would then do the garrison instead the AI cannot use the garrison units anymore. Therefore, the garrison units can a bonus for city defense so that the AI uses them.

This way, the AI now explores in the early game and can handle its troops better.
..\ctp2_code\ai\strategy\goals\Goal.cpp
..\ctp2_code\gs\newdb\goal.cdb
..\ctp2_data\default\aidata\Goals.txt