Make the AI not trying to load transports in cities without room
The AI may load transports in cities, however this only works if there is enough room for the units to load the transporter and the defenders of the city. Now if a transporter is in a city without room the AI will move it out. And won't move it again into the city. This way units won't wait forever for going on board.
..\ctp2_code\ai\strategy\agents\agent.cpp
..\ctp2_code\ai\strategy\agents\agent.h
..\ctp2_code\ai\strategy\goals\Goal.cpp
..\ctp2_code\ai\strategy\goals\Goal.h
..\ctp2_code\robot\pathing\robotastar2.cpp
..\ctp2_code\robot\pathing\robotastar2.h