Improve AI unit selection for siege
The AI now prioritizes the army agents on their absolute strength instead of their absolute strength, this has the advantage that the AI does not move armies from long distances to the target. However a bonus is added to transporters so that it prefers them. This has the advantage that it won't select another army once it has loaded a transporter.

The AI now selects armies that can beach assault, if it wants to conquer a city on a one tile island.
..\ctp2_code\ai\strategy\goals\Goal.cpp
..\ctp2_code\gs\gameobj\ArmyData.cpp
..\ctp2_code\gs\gameobj\ArmyData.h