Fix endless loop in path finding introduced by e9876bd
e9876bd made the entry costs for the A-star algorithm asymmetrical. That means depending where you are coming from the entry costs differ. Originally, the algorithm assumed that the costs of entering a tile are the same from all directions. However, if you land on a tile next to enemy units, these units can bombard you, but if you are one more tile away, they can't without moving. Thus there are some additional costs, not included in the terrain costs and they do not exit if the army lands a tile further away, especially if it then can directly attack those units.

To fix the problem the Astar algorithm does not consider paths that would make the current node a parent of its parent. In principle that is a problem that was always hidden in the implementation but never showed up.

..\ctp2_code\robot\pathing\astar.cpp