Bias the AI more to conquer weak enemies
The small target empire bonus and the weak enemy bonus is now also a applied if the enemy is not connected to the attacker. E.g. connected by a road or short or next to the attacker territory. The weakest enemy bonus is now applied to all enemies that are best for hot war.
..\ctp2_code\ai\strategy\goals\Goal.cpp

Improved debug reporting
..\ctp2_code\ai\strategy\goals\Goal.cpp
..\ctp2_code\ai\strategy\scheduler\scheduler.cpp

Make the AI prefer cities to conquer that are less threatened. And increase the city conquest boni for weak and small empires.
..\ctp2_data\default\aidata\Goals.txt

This targets the AI to one city civs. Those civs are created when a city declares independence due to a revolution and if the settings are so that this creates no civs instead of Barbarian cities. Inciting revolutions is a very effective strategy to cripple and destroy an empire. Of course such cities should be reconquered.