Fix include order of direction.h
direction.h must be included via MapPoint.h, because MapPoint.h defines _SMALL_MAPPOINTS. However, if _SMALL_MAPPOINTS is not defined then UP and DOWN are also defined which would also be enumerated in the following code:

	MapPoint neighbor;
	for(sint16 dir = 0; dir < NOWHERE; ++dir)
	{
		if(m_pos.GetNeighborPosition(static_cast<WORLD_DIRECTION>(dir), neighbor))

However, in case of UP or DOWN m_pos == neighbor. However, the code is not supposed to check the center. Therefore fixing the include order fixes, that the AI uses units for beach assault if the target city is not a one tile island but the access to it is blocked by a foreign unit.
..\ctp2_code\gs\gameobj\UnitData.cpp

Fix the include order of direction.h
..\ctp2_code\gfx\tilesys\TileDrawRoad.h
..\ctp2_code\gfx\tilesys\tiledmap.h
..\ctp2_code\gfx\tilesys\tiledrawroad.cpp
..\ctp2_code\gs\gameobj\CityInfluenceIterator.cpp
..\ctp2_code\gs\gameobj\CityRadius.cpp
..\ctp2_code\gs\gameobj\TerrImproveData.cpp
..\ctp2_code\gs\gameobj\Unit.h
..\ctp2_code\gs\gameobj\Wormhole.cpp
..\ctp2_code\gs\gameobj\armyevent.cpp
..\ctp2_code\gs\gameobj\player.h
..\ctp2_code\gs\world\MapPoint.cpp
..\ctp2_code\gs\world\WorldDistance.cpp
..\ctp2_code\gs\world\WrldCont.cpp
..\ctp2_code\gs\world\worldevent.cpp
..\ctp2_code\robot\pathing\Path.h