Make the AI ready to really if all units are on the target continent or being transported
The AI is only ready to rally if all units are on the target continent or being transported, previously it used the condition ocean continent for being transported, now it used the condition if the agent has cargo if it is on anther land continent then the target continent.

This way, the AI does not move a sub army to the target even so at hasn't grouped it with the other units.
..\ctp2_code\ai\strategy\goals\Goal.cpp