Add the move order events by the AI to the tail of the event queue
If the AI adds move orders to the tail of the event queue then it makes sure that the best stack of an army moves to a city for siege than the second best stack, in case the siege needs more than 12 units. This has the advantage that the best stack causes as much damage as possible so that it is easier for the second stack to win.

Note that ArmyData::PerformOrderHere is called from Goal::FollowPathToTask and already inserts the move order at tail for moving cargo. This is now extended to all types.
..\ctp2_code\gs\gameobj\ArmyData.cpp